+++ service/src/quiz/pages/QuizCreatePage/components/QuizCreateInputList/index.tsx
@@ -1,4 +1,4 @@
-import { Button, Calendar, DropDown, Text, TimePicker, Upload, useModal } from '@depromeet/toks-components';
+import { Button, Calendar, DropDown, TimePicker, Upload } from '@depromeet/toks-components';
 import { Flex, Spacing } from '@toss/emotion-utils';
 import { sub } from 'date-fns';
 import { ComponentProps, useRef } from 'react';
@@ -6,6 +6,7 @@ import { Control, FieldValues, UseFormRegister, UseFormReset, UseFormSetValue, U
 
 import { QUIZ_LIMIT_TIME } from 'quiz/pages/QuizCreatePage/constants';
 
+import { useConfirmModal } from '../../hooks/useConfirmModal';
 import { QuizCreateForm } from '../../types';
 
 interface QuizCreateInputListProps {
@@ -18,39 +19,6 @@ interface QuizCreateInputListProps {
   className?: string;
   reset: UseFormReset<QuizCreateForm>;
 }
-
-const useConfirmModal = (confirmCallback: () => void) => {
-  const { openModal, close } = useModal();
-  const openConfirmModal = async () => {
-    await openModal({
-      children: (
-        <Flex.Center direction="column">
-          <Text variant="title03" as="h3" css={{ margin: 0 }}>
-            작성 중인 퀴즈가 있어요.
-          </Text>
-          <Text variant="title03" as="h3" css={{ margin: 0 }}>
-            다시 만드시겠어요?
-          </Text>
-          <Text variant="body01" as="span" css={{ marginTop: '8px' }}>
-            다시 만들 경우, 지금까지 작성한 퀴즈 내용는 다시 복구할 수 없어요.
-          </Text>
-          <Spacing size={56} />
-        </Flex.Center>
-      ),
-      type: 'confirm',
-      buttonText: '다시 만들기',
-      onConfirm: () => {
-        confirmCallback();
-        close();
-      },
-      onCancel: () => {
-        close();
-      },
-    });
-  };
-
-  return { openConfirmModal, close };
-};
 
 export const QuizCreateInputList = ({
   register,
